refactor state_keeper tests
